refactoring to jsx
