+++ src/components/common/button/Button.jsx
@@ -1,18 +1,8 @@
-// Button.tsx
+// Button.jsx
 import React from 'react';
 import './Button.modules.scss';
 
-export interface ButtonProps {
-  label: string;
-  onClick?: () => void;
-  type?: 'button' | 'submit';
-  variant?: 'primary' | 'secondary' | 'danger';
-  size?: 'small' | 'medium' | 'large';
-  disabled?: boolean;
-  icon?: React.ReactNode; // Ensure icon is of type React.ReactNode
-}
-
-const Button: React.FC<ButtonProps> = ({
+const Button = ({
   label,
   onClick,
   type = 'button',
