+++ src/components/common/layout/Layout.jsx
@@ -1,11 +1,7 @@
 import React from 'react';
 import Header from '../../header/Header';
 
-interface LayoutProps {
-  children: React.ReactNode;
-}
-
-const Layout: React.FC<LayoutProps> = ({ children }) => {
+const Layout = ({ children }) => {
   return (
     <>
       <Header />
