+++ src/components/common/modal/Modal.jsx
@@ -1,13 +1,7 @@
 import React from 'react';
 import './Modal.module.scss';
 
-export interface ModalProps {
-  visible: boolean;
-  onClose: () => void;
-  children: React.ReactNode;
-}
-
-const Modal: React.FC<ModalProps> = ({ visible, onClose, children }) => {
+const Modal = ({ visible, onClose, children }) => {
   if (!visible) return null;
 
   return (
