+++ src/components/header/Logo.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-const Logo: React.FC = () => {
+const Logo = () => {
   return (
     <div style={{ width: '50px', height: '50px', backgroundColor: '#000', borderRadius: '50%' }}>
-      {/* Здесь можно вставить картинку или текст */}
+   
     </div>
   );
 };
